+++ Source/ZombieGame/Core/ZombieGameInstance.cpp
@@ -1,0 +1,9 @@
+#include "ZombieGameInstance.h"
+#include "../ZombieGame.h"
+
+void UZombieGameInstance::Init()
+{
+	Super::Init();
+
+	UE_LOG(LogZombieGame, Log, TEXT("ZombieGameInstance initialized"));
+}