+++ Source/ZombieGame/ZombieGame.cpp
@@ -1,4 +1,6 @@
 #include "ZombieGame.h"
 #include "Modules/ModuleManager.h"
 
+DEFINE_LOG_CATEGORY(LogZombieGame);
+
 IMPLEMENT_PRIMARY_GAME_MODULE(FDefaultGameModuleImpl, ZombieGame, "ZombieGame");