+++ Source/ZombieGame/ZombieGame.h
@@ -1,3 +1,5 @@
 #pragma once
 
 #include "CoreMinimal.h"
+
+ZOMBIEGAME_API DECLARE_LOG_CATEGORY_EXTERN(LogZombieGame, Log, All);